requirements.txt
neo4j
redisgraph
gremlinpython
